Do not re-compute the indexes, we can get them from unique

+++ src/reduce_cpu.cpp
@@ -6,7 +6,7 @@ template <typename scalar_t>
 void reduce_forwad_cpu_kernel(
     scalar_t* reduced,
     const scalar_t* full,
-    const int32_t* mapping,
+    const int64_t* mapping,
     int32_t n_samples,
     int32_t other_sizes
 ) {
@@ -35,7 +35,7 @@ void reduce_forward_cpu(
         reduce_forwad_cpu_kernel(
             reduced.data_ptr<scalar_t>(),
             full.data_ptr<scalar_t>(),
-            mapping.data_ptr<int32_t>(),
+            mapping.data_ptr<int64_t>(),
             n_samples,
             other_sizes
         );
@@ -46,7 +46,7 @@ template <typename scalar_t>
 void reduce_backward_cpu_kernel(
     scalar_t* full,
     const scalar_t* reduced,
-    const int32_t* mapping,
+    const int64_t* mapping,
     int32_t n_samples,
     int32_t other_sizes
 ) {
@@ -78,7 +78,7 @@ void reduce_backward_cpu(
         reduce_backward_cpu_kernel(
             full.data_ptr<scalar_t>(),
             reduced.data_ptr<scalar_t>(),
-            mapping.data_ptr<int32_t>(),
+            mapping.data_ptr<int64_t>(),
             n_samples,
             other_sizes
         );
